Ship_API/requirements.txt
# API 클라이언트
aiohttp>=3.9.0
asyncio-mqtt>=0.16.0

# 데이터 처리
pandas>=2.0.0
polars>=0.20.0

# Google API
google-auth>=2.25.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.110.0

# 날짜/시간 처리
pytz>=2023.3
holidays>=0.37

# 설정 및 환경변수
python-dotenv>=1.0.0

# 로깅
structlog>=23.2.0

# 기타 유틸리티
pathlib2>=2.3.7
typing-extensions>=4.8.0
